fetch data + from category

+++ admin-portal/src/app/(main)/categories/_utils/actions.tsx
@@ -2,14 +2,16 @@
 
 import { revalidateTag } from "next/cache";
 
-import { getAccessToken } from "@/lib/auth";
-import { ICustomError } from "@/utils/global-types";
 import { ICategory, ICategoryTree } from "@/app/(main)/categories/_utils/types";
-import { fetchData } from "@/utils/fetch-data";
+import { getAccessToken } from "@/lib/auth";
 import { CustomClassErrorApi } from "@/lib/exceptions";
-import { redirect } from "next/navigation";
+import { CategorySchema } from "@/schemas";
+import { fetchData } from "@/utils/fetch-data";
+import { ICustomError } from "@/utils/global-types";
 import { routePaths } from "@/utils/route-paths";
 import { statusCodeApi } from "@/utils/status-code-api";
+import { redirect } from "next/navigation";
+import { z } from "zod";
 
 const tag_revalidate_categories_list = "categories";
 
@@ -42,53 +44,79 @@ export async function getCategoryById(idCategory: string) {
 }
 
 export async function createCategory(
-  { id: parent }: { id?: string },
-  formData: FormData
+  id: string | undefined,
+  input: z.infer<typeof CategorySchema>,
 ) {
-  if (parent) formData.append("parent", parent);
-  console.log(formData);
+  const validatedFields = CategorySchema.safeParse(input);
+  if (!validatedFields.success) {
+    const error: ICustomError = { message: "Fields are invalid" };
+    return error;
+  }
+  const accessToken = await getAccessToken();
+  if (!accessToken) {
+    const error: ICustomError = { message: "You're unauthorized?" };
+    return error;
+  }
+
   const res = await fetchData<{ category: ICategory }>({
     url: "/category",
     method: "POST",
-    isProtected: true,
-    body: formData,
+    accessToken: accessToken,
+    body: { ...validatedFields.data, parent: id },
   });
   if (!res.success) {
     console.error(res);
-    return res;
+    return res.error;
   }
   revalidateTag(tag_revalidate_categories_list);
   redirect(routePaths.categories.path);
 }
 
 export async function updateCategory(
-  { id: idCategory }: { id: string },
-  formData: FormData
+  id: string,
+  input: z.infer<typeof CategorySchema>,
 ) {
+  const validatedFields = CategorySchema.safeParse(input);
+  if (!validatedFields.success) {
+    const error: ICustomError = { message: "Fields are invalid" };
+    return error;
+  }
+  const accessToken = await getAccessToken();
+  if (!accessToken) {
+    const error: ICustomError = { message: "You're unauthorized?" };
+    return error;
+  }
+  console.log({ validatedFields });
+
   const res = await fetchData<{ category: ICategory }>({
-    url: `/category/${idCategory}`,
+    url: `/category/${id}`,
     method: "PUT",
-    isProtected: true,
-    body: formData,
+    accessToken: accessToken,
+    body: validatedFields.data,
   });
   if (!res.success) {
     console.error(res);
-    return res;
+    return res.error;
   }
   revalidateTag(tag_revalidate_categories_list);
-  redirect(routePaths.categories.path);
+  redirect(routePaths.categories.children.category.path(id));
 }
 
 export async function deleteCategory(idCategory: string) {
+  const accessToken = await getAccessToken();
+  if (!accessToken) {
+    const error: ICustomError = { message: "You're unauthorized?" };
+    return error;
+  }
   const res = await fetchData<{ idDeletedCategory: string }>({
     url: `/category/${idCategory}`,
     method: "DELETE",
-    isProtected: true,
+    accessToken: accessToken,
   });
   if (!res.success) {
     console.error(res);
-    return res;
+    return res.error;
   }
   revalidateTag(tag_revalidate_categories_list);
-  return res;
+  redirect(routePaths.categories.path);
 }
